* origin/develop-1.10:
  edits to Connection Manager junit test
  Connection Manager junit test and formatting
  Removed parent and use shared parent pom.
  fix pom.xml for several unit test and include in either core or rcp pom.xml to maven verify stage to run
  upgrade tycho version to 0.18.1 from 0.18.0 to support maven 3.1+
  remove empty  gov.redhawk.testSuite plug-in
  remove empty tests/gov.redhawk.feature.tests plug-in
  add copyright, license and checkstyle to new gov.redhawk.bulkio.util.tests plug-in
  refs #2096 unit test case for BulkIOType.createServant
  refs #2139 workaround BooleanFieldEditor firing a property change event with String vs Boolean
  refs #2139 override store() and loadDefault() instead of doStore() and doLoadDefault() as it still cleared out the read-only field when OK is clicked
  refs #2139 ignore doLoadDefaults() and doStore() for read-only field on pref pages changed pref default for BulkIO remove on EOS from false to true to match the default of true in BulkIONxmBlockSettings
  refs #2133 proprogate sample rate from advanced Port Plot wizard all the way down change sample rate field from Double (double) to Integer (int) as it should always be integer and there was mismatch between double value in settings vs Integer field in PrefPage so that the value was not displayed properly when set.

+++ plugins/gov.redhawk.ui.port.nxmplot/src/gov/redhawk/ui/port/nxmblocks/BulkIONxmBlockSettings.java
@@ -31,8 +31,8 @@ public class BulkIONxmBlockSettings implements Cloneable {
 
 	/** custom connection ID to use. */
 	private String connectionID;
-	/** null to use default from StreamSRI. */
-	private Double sampleRate;
+	/** null (or zero) to use default from StreamSRI. */
+	private Integer sampleRate;
 	/** true to block pushPacket when downstream Command (e.g. Plot) cannot keep up, false to drop packets in this scenario. */
 	private boolean blocking;
 	/** null to use default pipe size from NeXtMidas (128K) */
@@ -84,14 +84,14 @@ public class BulkIONxmBlockSettings implements Cloneable {
 	/**
 	 * @return the current sample rate (null to use default)
 	 */
-	public Double getSampleRate() {
+	public Integer getSampleRate() {
 		return sampleRate;
 	}
 
 	/**
 	 * @param sampleRate the sampleRate to set (null to use default)
 	 */
-	public void setSampleRate(Double sampleRate) {
+	public void setSampleRate(Integer sampleRate) {
 		this.sampleRate = sampleRate;
 	}
